Remove test class duplicates
